Move protocol handling into its own folder
The Windows mail-client registration and its registry text helpers sit beside the mailto and meru:// handlers as protocol/, the way updater/ and gmail/ hold their own files. `lib/deep-link` stays where it is: the menu and context menu use it too.
