+++ packages/app/index.ts
@@ -16,15 +16,9 @@ import {
 import { ipc } from "@/ipc";
 import { initLinuxWindowControls } from "@/lib/linux";
 import { log } from "@/lib/log";
-import { registerWindowsMailClient } from "@/lib/windows-mail-client";
 import { licenseKey } from "@/license-key";
 import { main } from "@/main";
 import { appMenu } from "@/menu";
-import { theme } from "@/theme";
-import { appTray } from "@/tray";
-import { appUpdater } from "@/updater";
-import { doNotDisturb } from "./do-not-disturb";
-import { isMeruUrl } from "./lib/deep-link";
 import {
   findMailtoUrlArg,
   findMeruUrlArg,
@@ -34,7 +28,13 @@ import {
   PROCESS_MAILTO_URL_ARG,
   PROCESS_MERU_URL_ARG,
   setMeruProtocolClient,
-} from "./protocol";
+} from "@/protocol";
+import { registerWindowsMailClient } from "@/protocol/windows-mail-client";
+import { theme } from "@/theme";
+import { appTray } from "@/tray";
+import { appUpdater } from "@/updater";
+import { doNotDisturb } from "./do-not-disturb";
+import { isMeruUrl } from "./lib/deep-link";
 import { spellchecker } from "./spellchecker";
 import { trial } from "./trial";
